ACPI: move to consortium source (#201)
* ACPI: move to the standard acpica, not the linux one.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: get ACPICA to build.

This is from Intel. We can use the Intel source straight since we build as a library and restrictions
are somewhat less applicable.

This is from ACPI commit f39a732d2a0575b8a28b3d59f1836c7a3e4aa9cc

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPICA: Add tables.

Just a matter of adding a build file.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: remove unused platform/ (new one is in acpia/)

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: create build.json files for all components.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: Get user mode to build.

Kernel is temporarily disabled.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* acpi: get coreboot.h back.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: add new include files.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: start building the kernel with ACPICA

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: almost compiling.

And the init function in devacpi is calling a global init function.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: more attempts to get it to compile

It doesn't help that the ACPICA guys redefined a bunch of standard things.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: it *almost* now all compiles.

note we had to force an undef of _LINUX and __linux__.

Now we need to define AcpiOsInitialize, once we know what it does.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: compiles and builds a working kernel/userland.

As a test, devacpi::acpiinit, calls AcpiOsInitialize, which we also provide.

Many of the type names and declarations in ACPICA do not
follow the Plan 9 standard. We hope to wall most of the
usages behind the wall of devacpi. There's a lot of value
in not changing them, as one can then read the relevant
docs and see the same names in the doc as in our code.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: start calling functions.

Now the fun begins. There's all kinds of functions missing we'll need to provide.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: implement some functions.

There are others we still have no idea how to implement. Also,
the Flags to the lock stuff are unknown.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: set up the simpler functions.

I don't understand the harder ones.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: now builds.

I had to take the disassembler out of the kernel, but I am not
sure we even want it there.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* ACPI: re enable SMP

It was off while we worked things out, now it goes on.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* acpi: make acpi enabled by a flag.

This makes testing much easier.

At this point, I think the first pass of integrating ACPICA is done,
and things appear to work. I suspect if we want we can take this patch series
in. You may want to test it.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

* acpi: use the acpica library for cache utilities

That's what most other OSes do.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>