ACPI: run aml (it fails) to set _PIC to 1
This runs the _PIC code, which fails. When it works, it's the next step
in making ACPI work.

Also, if enableacpi is set in the boot args, main.c will do this:
enableacpi = acpiinit();

If acpiinit succeeds, enableacpi will be set, and we won't run
mpsinit().

This lets me slowly bring acpi to life, and lets others test the work,
all while still allowing the kernel to boot. This will make ACPI bringup
much easier, since we never break booting.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>