fixfmt: fix all usages of %u so that u is a type, not a modifier
The rest of the world uses %u to mean unsigned decimal. Plan 9
uses it as a modifier, which is arguably better, but what can we do?
You don't always get to pick the standard.

This change modifies libc/fmt so that %u is unsigned decimal, and
modifies most uses of it so the kernel almost boots. Almost. Obviously
we didn't get them all yet.

It hangs in ipconfig.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>