Fix GO9PRUN
Use the HARVEY variable so that we don't have to be any
particular place to run it correctly.

Add a shebang (#!/bin/bash) and set -e so some errors
get caught.

Add a port forward for 5356 so we don't get ridiculous delays on
start. CPU servers don't run auth servers anyway, but this cpu
server would hang on the auth server.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>