slice: get slice functions to compile.
They are now part of libc and can be used in kernel
or user mode. Their immediate use is in ACPI.

We've also adopting more Plan 9 style conventions:
no _ in names, use CamelCase
The name is now PtrSlice, as Slice was taken
use typedefs, not structs.
use the exiting reallocarray in our libc

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>