realloc: handle zero-length arrays.
We saw a problem in devacpi.c when realloc was called
to create a zero-length array: it returned a nil pointer,
not a pointer to a zero-length array, which confused the caller.

Consider the difference between:
char *s = "";
char *s = nil;

The first is a pointer to a zero-length string, the second
a NULL pointer. We want realloc to implement the first,
rather than the second.

I realize the comment is a bit long but this is an easy thing for
someone to come along later and "optimize" -- i.e., break again.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>