ACPI: add a program to set interrupts
If we're going to parse AML, that means interrupt setup
too. This will mean a program is required to start up
to configure PCI interrupts, but it's a lot safer.

Once we are
there we can work out interrupt settings with this program
we'll need to add a file to devacpi to let a user mode program provide
information so interrupts we could not set up can now be set up.

So this is going to require some plumbing but doing it in user mode
is a good thing. This is a prototype I hope to use on Akaros too.

This almost works, but dies trying to load the tables for some weird
reason.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>