ACPI: in the pci structs, track the "path"
The path to a device goes through bridges, and it's often useful
to know what it is (e.g. for ACPI, but for other things too).

As we scan the tree, track the path in a standard form
bus.dev.fn[/bus.dev.fn]* so that we can split it on the
/ later and easily create file names from these
paths so we can enumerate resources.

The pci comnmand correctly shows the path.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>