ACPI: a bit further
We're still not getting interrupts on the other side of a
PCI-PCI bridge set up right; I may know why.

This is an intermediate commit. It doesn't break any hardware
or qemu for me. In fact on qemu it seems to do the right thing.

To test on your system:

boot
bind a '#Z' /dev # this is the new name for ACPI; the other thing is hard to type

make sure '#$' and '#P' are bound

Pick a device to test and a pin. For example, on vmware, the IGBE is at
2:0.0 and it uses PIN B.
Run
acpi/irq 2 0 1 # i.e. bus 2 dev 0 PIN B

It will output lots of stuff, and then tell you what you can echo into /dev/irqmap

irqmap takes this:
segment bus dev fn gsi
and does its best to enable it.

Most people won't understand how this all fits together, I'm hoping people want to learn.
Then fix my bugs.

There are other little changes. pci now prints out the "path" of a device, i.e. the set of
bridges it is behind. There is a file programs can read which will return all the devices
where interrupts could not be set up.

The intent is that AML needed to set up devices is run in user mode; information for the AML
is provided by the kernel, and a program can write the needed info to /dev/irqmap to finish setup.

This means no AML interpreter in the kernel, which I think is a good thing.

So there's lots here, sorry.

I have to slow down on this for a while, I have two talks to finish working on.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>