ACPI: manage ACPI shared interrupts.
This change introduces a new arch file, irqenable.

We also modify the irqmap operation so that when a new interrupt is
mapped, i.e. new entry created in an IOAPIC RDT,
it is not enabled. This lets us set up all the
ACPI-created interrupts from the user level interpreter,
and then enable them later. This setup and enable
sequence is required in the case of shared interrupt
lines.

We were seeing the following problem: two NICS on VMWARE
had the same global system interrupt (GSI). When we set
up the first interrupt, and enabled it, we'd get an interrupt
storm as the second NIC could now issue an interrupt,
and the interrupt handler could not handle it, as the
second NIC was not set up: not knowing what to do, it just
returned. Interrupts on PCI are level,
not edge, so until the second NIC interrupt was handled,
we'd immediately drop back into the interrupt handler,
which would continue to do nothing.

The solution is to mirror what we do in the kernel: set
up the interrupts and, once done, enable them.

This #P/irqenable device, on a write, unconditionally
enables all IOAPIC RDT entries that have a non-zero
vector. This is kind of inflexible and we may want to add
the ability later to allow enabling interrupts based
on [bus, device] pairs as in irqmap.

I doubt we'll ever need that flexibility but who knows.

This may, finally, give us working interrupts on
ACPI-only platforms. Now we can start to clean up
the code.

We have also avoided jamming the tens of thousands of lines of code
that is the ACPICA code base into our kernel. That's good, because ACPICA
is not.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>