ACPI: create a file in #P for acpimem.
The way I did ACPI memory was not right. Qraw should be for the
raw data for one ACPI table and physical addresses should not be
used as offsets in it.

Create a new arch file, acpimem, that works like the old realmodemem
file, in that you can only read parts of it that correspond to the ACPI
tables.

This is the first in a sequence of changes making it possible to use the ACPICA tools
(e.g. acpidump) for debugging. I'm still trying to track down what broke in the VMWare
Fusion world and the story just keeps getting worse there (two different VMs on one
mac have very different ACPI behavior).

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>