Various bits of aarch64 code to support ARM64.
Implement (or copy from another architecture) most of the
bits required for building libraries and applications for
aarch64.  The system more or less builds now.

I'm sure there are some bugs in this code: see, for example,
libc/aarch64/main9.S: this is the "get to build" stage.  The
next stage is "make work."

Signed-off-by: Dan Cross <cross@gajendra.net>