getcallerpc/getcallstack: move common code to libc
Dan brought in nice standard code but pointed out it was
being replicated. This change allows us to use the common
code for cases it can be used. For special cases,
one can still create special versions in the libc/$ARCH
directory. Which one is used depends on libc/$ARCH/build.json.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>