Move devether.c to IP stack.
devether is the glue that ties the IP stack
to ethernet devices; it's essentially a kernel
API with a filesystem interface.

There is some suspect, possibly 386-only code
in the probe routine having to do with IRQ
selection, but if it turns out to be a problem
on other architectures, we can fix that.

Signed-off-by: Dan Cross <cross@gajendra.net>