Fix BUILD files for the new build tool
Also add a bit of info to bootstrap.sh.

The new build tool is still not useable but at least people can see
how nice it's going to be.

I can't see how to make sys/src/cmd/BUILD build
all binaries. I also don't see why I have to list
all the installs AND all the builds -- why isn't
that just implicit? If I build something I want to install
it. At least that's how plan 9 did it.

And I note that the source has a name like "bin/ed"
in the install descriptor, and the destination
a name like amd64/bin/ed -- if this tool is harvey
specific (it is) it ought to know about $ARCH.

I'm also not sure why harvey_binary is special.
Does the build tool really need to be specialized
to harvey? If so, why not delete all non-harvey items in
it? If not, well, why is the name harvey_binary there?
This seems wrong.

And if I'm going to have a name like harvey_bin why can't it
be smarter about harvey rules?
Why can't it just do everything needed the way
the current tool does? build it, put it in $ARCH/bin.

The non-verbose mode, while it's kind of cute with the display,
hides error messages. That's bad. We need to get rid of it and just
have the -v mode be the standard.

It doesn't emit errors that emacs can deal with. We need to fix the
error output. Not sure how to do this with a tool with such great
concurrency but we have to do something.

We need to get this tool into the harvey project and get it
working. We can all help.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>