riscv: adjust computation for mappings in PADDR
I'm working on a long note to put into the code,
but the short form here is that riscv memory starts
at 0x80000000, whereas on x86 it starts at 0. This
broke as assumption in PADDR about how to map kernel
virtual to physical. This is a simple fix that makes
it work again.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>