riscv: handle the fact that we can get 3 or 4 level page tables
This should work for now, but it's not nice. Basically if we're on a machine
with a 3 level pt we fake up the 4th level to make the software simpler. This basically
is hurl-inducing but gets us further in boot while we work out how we really
want to handle it.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>