riscv: add configstring parsing code
This is the initial commit, the code is not very nice.
I'm going to merge this into one file, get rid of the
totally unnecessary inlines, and clean up the mess.

This won't come close to compiling.

We require this code so we can find the timer
comparator register, which can only be found by looking
at the configstring. It will also allow us to find other neat
things such as the UART and # of harts.

The configstring will be passed from coreboot as the first
parameter. I'm trying to get the riscv community to settle
on this simple model, i.e. first parameter to kernel
entrypoint is the configstring, but they keep
fixating on doing it via a callback to firmware. Ugh.

It is from lowrisc-bbl 87588c4

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>