riscv: fix translation of pte to ppn in mmu.c
Unlike intel, the ppn in the pte can't be translated just by
blowing the low bits clear. It's a 4k ppn but it is only
shifted left 10 bits. It needs to have the low 10 bits
cleared and then get shifted left 2 bits.

It's really amazing this worked as long as it did
with this error in place.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>