riscv: more changes
And, because GNU LD sucks, we need ldscripts to make things
align correctly. And, we need to ignore the ELF alignment for
each program header, since GNU LD doesn't seem to let us set it.

I'm up to sysexecregs, and I have other things I really have to get doing,
so this is going to rest for a bit. I really need to get it upstreamed,
but will probably do it as a squash commit.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>