riscv: reconciliations and bug fixes
amd64/core.json now includes virtio, since that's portable
but not usable with riscv yet. mmutlbflush, as designed,
can't take a parameter on pretty much everything. We can
change it back if it ever makes sense, but its use in port/
had an amd64 dependency wired in. Elf64 has to be more relaxed
about alignment in ELF program headers, due to gcc issues.
syscallfmt needs to be more careful about not throwing an
error when a system call has invalid arguments.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>