strace command for new tracing
This is the second go around of tracing that I did for Akaros,
based on the earlier one I did for Plan 9. It's the best yet.

The tracer does not work. The processes don't have to block.
System calls appear in proper order.

TODO: see the commented out functions for release and free?
They need to be filled in else, over time, you leak data.
Any such leak would take weeks or months to be a problem,
so we can use this as is, with the annoying 'are we leaking'
message in place. Maybe some smart person can finish this up.

But, for now, it's far superior to ratrace, which I think we should
plan to remove soon.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>