riscv: use the SBI call to set mtime compare (mtimecmp)
Due to a bug in the RISCV simulator we were able to set
mtimecmp (still allowed) and also able to clear the
Supervisor Timer Interrupt Pending (STIP) bit. That bit
can only be cleared in M-mode but SPIKE let us do it.

We had to change timerset to do an mcall.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

Conflicts:
	sys/src/9/riscv/devarch.c