bldy: let bldy take a vacation for a bit
Bldy is a brilliant prototype but we want to rethink some aspects
of its implementation. Until we're sure how it's going to work
out, take it out of the test matrix (it's pushing our travis
times to 30 minutes) and don't get it or mention it in
bootstrap.sh. Do this in a way that lets us bring it back
when we're ready.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>