Cleaned up warnings in libfreetype
Mostly unused variables, or set-but-unused.

The unfortunate practice seems to be to define a variable,
initialize it to zero, and set it to one later.  Then, presumably,
someone could look at it with a debugger or something and see
that it had been set at some point (or not).  Perhaps this is
used as a rough signal to see that something has happened in a
function.

Unfortunately, an optimizing compiler will probably note that
the variable is set but unused and just remove it.  So we just
remove some of them ourselves.  Or whatever.

Signed-off-by: Dan Cross <cross@gajendra.net>