When writing to text segments, handle both SG_TEXT and SG_LOAD
When copying a text segment into a data segment for writing, setup permissions correctly.  Similarly when duplicating data as text.  These changes needed because SG_RONLY was replaced with SG_READ and SG_WRITE, and SG_LOAD seems to have been added as a replacement for SG_TEXT, but the work wasn’t completed.  We should either work out what was intended, or remove the change in the future.

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>