Implement single step - incomplete, but allows continuing
After a breakpoint is hit, if the user continues, the first packet is 's' - single step.  This implementation is incomplete, amd64 only, and is known to cause issues if the user deliberately single steps, but as it allows continuing after hitting a breakpoint, it's worth checking in to be improved later.

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>