Fix bug in libmach where mapped text segment didn't include stack
The format of /proc/pid/segment changed in commit 121208f.  Originally, if a segment was read-only, 'R' would be shown after the segment name.  In the case of the stack, this would be absent.  After that commit, there would always be a token displaying the segment permissions, therefore changing the format.  Also, due to the optional 'P' token, it is difficult to parse reliably.

The symptom was that any mapped data segment didn't include the stack.  This was because the change in format to segment meant when creating the map, we were reading the lower address for the stack, rather than the top.  This showed up when trying to view the stack in gdbserver -
 e.g. displaying backtrace.

This commit also combines the 'P' token with the permission section so we always know which tokens are where.  Segment doesn't seem to be pa
rsed anywhere else that I can find, so should be safe.

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>