Implement get_reg for gdbserver
Rather important for getting the stack pointer to step out of functions.
Also tidy the formatting of some error messages that showed up alongside the bug.

Seems the list of registers is duplicated in sys/src/9/amd64/dat.h
We should probably pick one place for this to avoid the duplication.

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>