UFS: Mounting the softdep code
This brought in a lot of scattered changes unfortunately.  I've disabled snapshots for now (NO_FFS_SNAPSHOT).  Another slight bit of ugliness is 'struct workhead'.  Need to keep the struct due to the way the LIST_HEAD macro works.  If I change that (yes, I tried) then it breaks somewhere else that calls the macro as so: LIST_HEAD(, snapdata)....

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>