Upgrading cputype
There's no need for a function and dev telling about number
of cores in the cpu, it should be part of the dev/cputype.
Now cputype can handle all the supported cpus in amd64/archamd64.c
For now seems nobody tried harvey in a machine without an AMD or Intel
processor, but who knows.
With this change, every supported cpu added to archamd64 will be showed
in dev/cputype.

TODO: some specific family and model info.

Signed-off-by: Álvaro Jurado <elbingmiss@gmail.com>