Avoid unnecessary retries in sdiahci
We get the retry because ci is cleared between the time it's set and we test it before sleeping.  We don't clear ci manually, but it can be cleared if the ST flag on p->cmd is cleared (and probably for other reasons too).  We clear the st flag at various points throughout the sdiahci code

This is a little hacky, but ensures we always sleep, which would always happen in the usual case anyway.

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>