Ensure libthread stack aligned to 16 bytes for amd64
Ensure we'll still be aligned to 16 bytes after return address is pushed onto the stack.  Important because otherwise we'll get a GP exception if we try to use some SIMD instructions with a stack that's not aligned to 16 bytes.

This could be seen by building with gcc and running usb/usbd without any USB hubs.  An error case pushed the code into threadsysfatal, which had been compiled into some MOVAPS SIMD instructions, which expect the stack to be aligned to 16 bytes,

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>