Renumber harvey system call numbers
Create a new struct member, SyscallBase, in sysconf.

This is a base from which to offset Harvey system calls.

In mksys.go, add sysconf.SyscallBase to the harvey system call numbers.

This allows us to easily define a base from which to offset Harvey
system calls.

The next step is to add a Plan9SystemCall element to sysconf.json,
then generate those system calls, then add the handler in the
kernel, which should be a pretty simple shim. It's been done
before for NxM, in the opposite direction, when we did Linux
system call emulation, so the code to switch out on system
call number and so on exists.

I'd suggest adding back Plan 9 binary support by, first, translating
plan 9 a.out format to ELF, then running it; that's likely easier than
putting plan 9 a.out format support back in the kernel (Plan 9 a.out
for amd64 was a terrible hack and no Plan 9 a.out was ever defined for
RISCV anyway).

As jmk used to say, "Plan 9 is a 32-bit system." That's true. Harvey
is much closer to being a 64-bit system then Plan 9 ever was, there's
not reason to bring back Plan 9 a.out format.

Signed-off-by: Ronald Minnich <rminnich@gmail.com>