The empire strikes back
Bring a.out format back. This is to help Go support.

We tried, oh how we tried, to get a harvey port of Go. That's a massive
job and nobody had the time. Better to try to run Plan 9 binaries.

We were thinking to build ELF files for Plan 9
from the Go compiler, but this won't happen.
I've talked to someone on the Go core team and Go is not going to be
flexible on this point.
We need to be able to run classic Plan 9 a.out binaries.

It's a shame we can't also convince Plan 9 go compiler to use the
standard ABI, but no luck.

This is step 2 in the process of supporting Plan 9 Go binaries.
The first was renumbering Harvey system calls. The next
is supporting old school Plan 9 system calls and on-stack args.

This now runs an empty Go plan 9 binary, and dies on a bad system call.
i.e.
package main
func main(){}

GOOS=plan9 go build main.go

run main in Harvey

this includes step 3 with copious debug prints, it runs a Go program.

The TOS had to be changed to match Plan 9 TOS, so Go would work.

Signed-off-by: Ronald Minnich <rminnich@gmail.com>