Fix seek for Plan 9 binaries
Plan 9 is a 32-bit kernel. This is reflected in the
seek that Go uses, in which the first param is a
uint64_t*.

if up->plan9 is set, check the system call number and,
if it is SEEK, modify behavior:
o use args[0] as a uint64_t* and verify with validaddr
o use args 1-x instead of 0-x
o store ar0.p into the pointer on return.

This fixes Go environment walking among other things,
and enabled the u-root rush shell to work.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>