Fix multiboot args
When I ported this to att assembly, and brought in coreboot code,
I unknowingly broke multiboot args. This fixes them.

The code is nasty, but assembly always is.

If at any point you are tempted to change this, you MUST test with
a multiboot boot and a normal boot.

Remove other hacks that also broke multiboot tables.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>