Fix a broken compare in 386/uarti8250.c
Here's a rare find: a bug that's been there forever.

The code wrote the Scr, then read it back. If the Scr is there, it reads
back the same. If there is no Scr, it reads back different.

Somehow the test got reversed, and for at least 5 years, harvey was not adding
the uarts for console.

It is hard to say how old this bug is. But it is old.

With this change, on boot, we finally see some uarts:

     0x0 0xffffffff ACPI            0x0 0xffffffff ACPI            0x0      0x0 ACPI
    0x60     0x60 keyb
    0x64     0x64 keyb
    0x70     0x71 rtc/nvr
   0x2b0    0x2df vga
   0x2f8    0x2ff COM2
   0x3c0    0x3da vga
   0x3f8    0x3ff COM1
   0x800    0x803 ACPI
   0x804    0x805 ACPI
   0x810    0x817 ACPI
   0x818    0x81b ACPI
   0xfff    0xfff dummy

Before, they did not appear.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>