uarti8250: make COM1 and COM2 poll, or any uart with poll or tbdf==-1
now that UARTs are showing up in #t again, any attempt to use them, e.g.
echo hi >> '#t/eia0'
will get a panic, as our interrupt support code no longer supports
ancient hardware with a tbdf of -1.

Note that, cleverly, uart interrupts are never even enabled until the
device is opened, which is why we did not see this problem earlier

We made this change I believe in the mod 2000s as part of the Fast-OS effort,
as we tried to drag the code forward from its 1990 base. It's a shame we
did not have code control back then.

i8250enable now checks to see if ctlr->tbdf is -1, or ctlr->poll is set;
if so, it disables interrupt enable.

The i8250poll code now runs, finally.

But what broke between 2011 and how such that serial consoles stopped working.
Therein lies a puzzle.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>