Fix broken iterator in uartpnp
This was a mess. If it ever worked, it was by accident.

At least now I understood why the Scr test was broken the
way it was. It was a half-hearted attempt to fix this broken
code. But it did not work.

This now correctly finds uarts on machines with 1 or 2 8250s,
qemu and the apu2 in this case.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>