Get uarti8250 working again; add a bit more debug info for boot
devuart will now tell you more about what uarts it found and why. This
is useful; the old boot path was silent. Silence is golden until you want
to know if you have uarts.

boot.c now includes an optional function, catstuff, that prints some info out pre-boot, e.g.
     0x0 0xffffffff ACPI            0x0 0xffffffff ACPI            0x0      0x0 ACPI
    0x60     0x60 keyb
    0x64     0x64 keyb
    0x70     0x71 rtc/nvr
   0x2b0    0x2df vga
   0x2f8    0x2ff COM2
   0x3c0    0x3da vga
   0x3f8    0x3ff COM1
   0x800    0x803 ACPI
   0x804    0x805 ACPI
   0x810    0x817 ACPI
   0x818    0x81b ACPI
   0xfff    0xfff dummy

          3           0                    0                    0 trap     #BP
          8           0                    0                    0 trap     #DF
         14           0                   31             38958038 trap     #PF
         15           0                    0                    0 trap     #15
         19           0                    0                    0 trap     #XM
         50          50                 1255              6247693 lapic    APIC timer
         65           1                    0                    0 ioapic   keyb
         73           0                    0                    0 msi      ether0
         81           0                    0                    0 msi      ether1
         89           0                    0                    0 msi      ether2
         97           9                    1            140323595 ioapic   acpi

This has proved to be immensely helpful on debugging.

Finally, since we can't do i8250 interrupts any more, and nobody knows what a modem is,
and serial is mostly 3 wire devices, rip all the ancient non working code out.

i8250 now runs purely as a clock-interrupt-driven device. We can change this back any time
if it's a problem.

But actual, functional, working, serial console is almost back for
the first time in many years.

In general, uart is a debug, little used, device. But when you need it, you really need it.
What we did on many systems was run a program called tty, which fed bytes to a shell.
We can try to find that (it was used on k10 and blue gene kernels) but for now this works.

So how long has uart been broken? I no longer recall.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>