get serial console finished.
It's a hack but it's working and will let us debug on atomic pi, apu2, and others.

The uart is enabled by forking a child in boot.c and having it open #t/eia0.

As long as that process is around the uart is polled. Kill it and polling stops.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>