devramfs: get it to work
devramfs violated a lot of rules about writing drivers, in its not quite
right use (or lack of) of error etc. It returned 0 when it needed to call error(),
and so on.

It also panicked the kernel.

This verion will let your write up to a 32MiB file and limits
writes to 8k. It pre-allocates an array to hold up to 4096 8K
blocks and allocates the blocks on demand. It mistakenly
does not track file size and needs to differentiate reading
past end and returning big blocks of zeros on read when there is no block.

So here you go, a few things you can consider fixing and learning from.

Needs work, but it's far more workable than what we had.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>