Add support for Intel Z170 chipset to sdiahci.c sdiahci can now identify the z170 chipset and its variants.
At least on my machine, it hangs when trying to initialise ahci port 5 during boot.
When calling iapnp in sdiahci (which then calls newctlr, then achiidle). It tries to dereference port->cmd as so: if((*p & Arun) == 0).
If that port is skipped, it's able to continue the boot process.

I think it's safe to check this in without a fix for the port issue as they seem to be separate problems.

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>